Make nightly job only responsible for updates
This switches the nightly update to using a personal access token, instead of the built-in GITHUB_TOKEN.

The GITHUB_TOKEN is special in that when it is used to push code, it does not trigger any other workflow that would normally happen on push.

Because of that limitation, we had turned the build workflow into a reusable workflow and called it explicitly from the nightly update.

But, that causes another limitation: reusable workflows cannot call other reusable workflows, which meant that we had to define the build jobs for all of the submodules twice: once here, and once in the submodule.

With this change, we can now make the submodule builds into reusable workflows and just call them with a single line from the mega-build here instead of having to define the jobs twice.