fix nightly update for real
I had switched to using the `submodules: true` option on the github checkouts action,
but that does something fancy to reduce the download size, which has a side effect of breaking updates.
So, kinda the opposite of what we need here.

I rolled it back to a simple `git submodule init` and now it appears to be working again.